+++ TODO
@@ -1,6 +1,5 @@
 By priority:
 
-- tiled matrix multiplication
 - safe multiplication kernel with checksums and error correction/detection
 - async copy while computing buffering schemes:
     - AB->C (default)
