--- include/main.h
@@ -1,12 +1,0 @@
-#ifndef __MAIN_H_
-
-#define __MAIN_H_
-
-#include "utils/alloc_matrix.h"
-#include "utils/check_product.h"
-#include "cli.h"
-#include "kernels.cuh"
-#include "parameters.h"
-#include "utils/print_matrix.h"
-
-#endif
