+++ include/programs.h
@@ -1,0 +1,4 @@
+namespace programs
+{
+    int gigacheck(int, int, int);
+}
